Fixed issues surrounding login and signup components not being displayed properly
